backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-docx>=0.8.11
pandas>=2.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
requests>=2.31.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
python-multipart>=0.0.6